+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 2.2.15
+* Factories can be marked as catalog item to support semantic copy. 
+
+```java
+public static class FactoryCatalogItem extends SimpleFactoryBase  {
+    public FactoryCatalogItem {
+        this.config().markAsCatalogItem();
+    }
+}
+```
+
 # 2.2.16
 
 ### BREAKING CHANGES